Add optional Jitter to long poll to help smooth out load over time.
