+++ internal/pkg/config/config_test.go
@@ -111,6 +111,7 @@ func TestConfig(t *testing.T) {
 								Write:            5 * time.Second,
 								CheckinTimestamp: 30 * time.Second,
 								CheckinLongPoll:  5 * time.Minute,
+								CheckinJitter:    30 * time.Second,
 							},
 							Profiler: ServerProfiler{
 								Enabled: false,
